requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.1
click==8.1.7
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.5
FastAPI-SQLAlchemy==0.2.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
jose==1.0.0
Mako==1.3.6
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.10.1
pydantic_core==2.27.1
PyJWT==2.10.1
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
types-passlib==1.7.7.20240819
typing_extensions==4.12.2
uvicorn==0.32.1
